Initial implementation with basic functionality.
This is the initial implementation, and thus limited
in fairly many important ways. Right now it particularly
needs attention to the support of different payload
signature schemes as it does not handle streaming ones
at the moment, it also does not yet with the iRODS S3
resource as a client.