[#50] Refactor authentication logic for iRODS 4.2.
Before this commit, the server would do the following when a request was
received.

  1. Create a rodsadmin connection based on the credentials in the
     config file. The connection is not proxied at this time.
  2. Calculate the signature, compare it against the received signature.
  3. If the signatures match, rc_switch_user would be called on the
     connection, turning it into a proxied connection.

That worked, but only for 4.3.1 servers. With this commit, step 3 is
replaced with the following.

  3. If the signatures match, the iRODS username is returned.

The caller now checks if they received a string. If yes, the connection
that was originally created is replaced with a new proxied connection.

While slower, this solution is the only way to support 4.2 given 4.2
does not support the rc_switch_user API endpoint.