[77] Refactor project to not use C++20 coroutines.
This commit lays the foundation for moving away from C++20 coroutines.
This commit removes all S3-specific logic. That means the features which
shipped in v0.1.0 of this project will need to be ported to the new
design.

The changes in this commit were ported from HTTP API v0.2.0. All iRODS
4.3.1 features used by the HTTP API are included.

Notable features are:

    - 4.2 compatibility
    - configuration validation
    - connection pooling
    - secure communication with iRODS