[#114] implemented multipart enhancement
Design:

1. When a new part upload is initiated, determine if all lower
numbered parts have been started. If so, you can determine the part
offset via the previous part sizes.  In this case stream the part
directly to iRODS with a seek to the offset and writes.

2. To accomplish the above, when a new part upload is initiated, save
the part size in a map that translates the upload_id's to a list of
parts numbers and sizes. The following is an example of this map:

{
  "1234abcd-1234-1234-123456789abc":
    { 0: 4096000,
      1: 4096000,
      4: 4096000 },
  "01234abc-0123-0123-0123456789ab":
    { 0: 5192000,
      3: 5192000 }
}

3. If the part offset is not known, write the bytes to a local part
file. When CompleteMultipartUpload is encountered, read all of these
local part files and stream these to iRODS. If there is no local part
file, that means that that part was streamed directly to iRODS and does
not need to be rewritten.

4. The first open to iRODS will remain open until
CompleteMultipartUpload is finished. This is done to make sure that the
replica_token does not update in the middle of writing parts to iRODS.
See the keep_dstream_open_flag flag in putobject.cpp.