[#157] Add GitHub Actions Workflow for building on Ubuntu
irods-runtime 4.3.1 and 4.3.2 use two different versions of clang/clang-runtime,
so we must install both in order to be able to build both in the same container.
Ubuntu 24 has no package for clang 13.0.0 (which is what 4.3.1 uses), so we cannot
use the same workflow definition for both until and unless 4.3.1 support is dropped.
Until then, we will split them up.

This commit also disables IRODS_BUILD_WITH_WERROR in the builder image script.