[#148] Enable running specific tests
This commit enables running specific tests by allowing the command of
the test-runner service container to be overridden with a list of test
suites, classes, and/or names. This is possible because the command is
simply passed as arguments to the entrypoint script, which uses the
arguments in its call to the unittest CLI. That means that all of the
features of the unittest CLI are available for use.